+++ CHANGELOG.md
@@ -1,14 +1,21 @@
+## 2023-12
+* Update Voidlinux url https://repo-default.voidlinux.org.
+* Add global use="modules-sign" for Gentoo.
+* makeopts on Gentoo use the value of `nproc`.
+* Correct install on Gentoo with systemd (update kernel name and more).
+* More rubocop style.
+
 ## 0.3.5, release 2022-11-21
 * Tested on a live Ubuntu 22.10.
 * Ensure `dracut.conf.d` exist before writing to it.
 * Display the version with `-v`, `--version`.
 
 ## 0.3.4, release 2022-10-10
-* Can work on a Live image of Voidlinux
+* Can work on a Live image of Voidlinux.
 
 ## 0.3.3, release 2022-10-01
 * Support disk with a sector size of 512.
-* Support vdx disk.
+* Support vdx disk (disk on virtualization).
 
 ## 0.3.0, release 2022-02-17
 * Gentoo with musl use an additional repo https://github.com/gentoo/musl.git.